In scrypt remove the check that N < 2^(128 * r / 8)
This bound is given in RFC 7914, and would imply that scrypt could not
make use of more than 16 MB of memory. This bound is contradicted by
the RFC itself, which includes tests which use up to 1 GB.

An errata was filed for the RFC in 2020 regarding this issue, but
never acted upon by the authors.

See #24650 for background

Co-Authored-By: Neil Horman <nhorman@openssl.org>